// client/src/types/productType.ts
export type productType = {
  _id?: string;
  // user: string;
  author: string;
  description?: string;
  image?: string;
  price: number;
  productId?: number;
  title: string;
  language: string;
  condition: string;
  location: string;
  year: number;
  pages: number;
  createdAt?: Date;
  updatedAt?: Date;
  __v?: number;
};